requirements.txt
imageio==2.13.5
matplotlib==3.4.2
ml_collections==0.1.0
numpy==1.20.3
pandas==1.2.4
Pillow==9.0.0
scipy==1.7.0
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tqdm==4.62.2
